refactor: create custom hook to hold home page logic
